标注工具_完整独立版/requirements_标注工具.txt
# 标注工具依赖库
Pillow>=9.0.0
